Moved discovery strategies to seperate file
